change: ログインの仕様変更
