+++ SakodaTrainingRSSReaderApp/SignUpViewController.swift
@@ -1,0 +1,15 @@
+//
+//  SignUpViewController.swift
+//  SakodaTrainingRSSReaderApp
+//
+//  Created by sako0602 on 2024/06/30.
+//
+
+import UIKit
+
+class SignUpViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
